Vitis-486 Use byte copying of argument data
This switches from 4 bytes copying to single byte copying in order
to allow for host application kernel arguments that are not 4 byte
aligned.  For example host application can set scalar arg of any size
including single byte.  When copying host data, ABR could occur if
assuming 4 byte aligned.

Retain xocl::kernel::argument rather than separate containers for
the various types of arguments which got out of hand.

This PR also removes dead code.

+++ src/runtime_src/xdp/profile/plugin/ocl/xocl_profile.cpp
@@ -37,7 +37,7 @@ static unsigned int
 get_cu_index(const xrt::run& run)
 {
   auto& cumask = xrt_core::kernel_int::get_cumask(run);
-  for (size_t bit = 0; bit < cumask.size(); ++bit)
+  for (unsigned int bit = 0; bit < cumask.size(); ++bit)
     if (cumask.test(bit))
       return bit;
   return 0;
